+++ src/NUnitFramework/tests/ExecutionHooks/TestAttributes/ActivateLongRunningBeforeTestHookAttribute.cs
@@ -4,7 +4,7 @@ using System;
 using System.Threading;
 using NUnit.Framework.Internal;
 
-namespace NUnit.Framework.Tests.ExecutionHooks.Common
+namespace NUnit.Framework.Tests.ExecutionHooks.TestAttributes
 {
     [AttributeUsage(AttributeTargets.Method)]
     internal sealed class ActivateLongRunningBeforeTestHookAttribute : ExecutionHookAttribute
