+++ src/NUnitFramework/tests/ExecutionHooks/TestAttributes/TestActionLoggingExecutionHooksAttribute.cs
@@ -3,7 +3,7 @@
 using System;
 using NUnit.Framework.Internal;
 
-namespace NUnit.Framework.Tests.ExecutionHooks.Common
+namespace NUnit.Framework.Tests.ExecutionHooks.TestAttributes
 {
     [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
     internal sealed class TestActionLoggingExecutionHooksAttribute : ExecutionHookAttribute
